+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [1.0.0] - 2018-06-04
+### Added
+- Rebuilt whole project with React.
+- Singit as submodule.
+- Dadgad-tuner as submodule.
+
+### Removed
+- Gulp.
+
 ## [0.7.2] - 2018-04-22
 ### Changed
 - Perlin project gets showed first.